requirements.txt
pypdf2
pillow
transformers
sentence-transformers
torch
scikit-learn
pymupdf
pyngrok
tqdm
